More hole filling

+++ src/main/java/hellofx/fheroes2/maps/Tiles.java
@@ -330,7 +330,7 @@ public class Tiles {
     }
 
     public void SetObject(int mp2Kind) {
-        mp2_object = (byte) mp2Kind;
+        mp2_object = (short) mp2Kind;
     }
 
     public int GetObject(boolean skip_hero) {
